supp: add multi-worker architecture with signal-safe shutdown and RPC timeouts
Replace the single-threaded request loop with a pool of worker threads
that handle concurrent TEE RPC requests. Register signal handlers for
SIGTERM and SIGINT with a safe shutdown sequence that drains pending
requests before joining workers. Add configurable timeouts to all
blocking RPC waits so the supplicant cannot hang indefinitely on a
stalled kernel RPC channel.

+++ reefs.c
@@ -4,10 +4,10 @@
  * FS supplicant for mbedtee-reefs
  */
 #include <fcntl.h>
-#include <math.h>
 #include <pthread.h>
 #include <sched.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -19,36 +19,73 @@
 
 #include "supp.h"
 
-//#define MSG printf
-#define MSG(...) do {} while (0)
-
 #define REEFS_PATH_MAX (1024)
 
-#define REEFS_PATH "/data/mbedtee/reefs"
+#define REEFS_PATH "/var/local/mbedtee/"
 
-static int reefs_mkdirs(const char *path, mode_t mode)
+#define REEFS_MAX_DIRS 128
+static DIR *reefs_dirs[REEFS_MAX_DIRS];
+static pthread_mutex_t reefs_dirs_lock = PTHREAD_MUTEX_INITIALIZER;
+static pthread_mutex_t reefs_meta_lock = PTHREAD_MUTEX_INITIALIZER;
+
+static int reefs_alloc_dir(DIR *d)
 {
-	int i, len = strlen(path), ret;
-	char dir[REEFS_PATH_MAX];
+	int i = 0;
+	pthread_mutex_lock(&reefs_dirs_lock);
+	for (i = 0; i < REEFS_MAX_DIRS; i++) {
+		if (!reefs_dirs[i]) {
+			reefs_dirs[i] = d;
+			pthread_mutex_unlock(&reefs_dirs_lock);
+			return i + 0x10000;
+		}
+	}
+	pthread_mutex_unlock(&reefs_dirs_lock);
+	return -EMFILE;
+}
 
-	if (len >= REEFS_PATH_MAX)
-		return -ENAMETOOLONG;
+static DIR *reefs_get_dir(int fd)
+{
+	DIR *d = NULL;
+	int idx = fd - 0x10000;
+	if (idx < 0 || idx >= REEFS_MAX_DIRS)
+		return NULL;
 
-	memcpy(dir, path, len);
-	dir[len] = 0;
+	pthread_mutex_lock(&reefs_dirs_lock);
+	d = reefs_dirs[idx];
+	pthread_mutex_unlock(&reefs_dirs_lock);
+	return d;
+}
+
+static int reefs_free_dir(int fd)
+{
+	int idx = fd - 0x10000;
+	if (idx < 0 || idx >= REEFS_MAX_DIRS)
+		return -EINVAL;
+
+	pthread_mutex_lock(&reefs_dirs_lock);
+	if (!reefs_dirs[idx]) {
+		pthread_mutex_unlock(&reefs_dirs_lock);
+		return -EBADF;
+	}
+	reefs_dirs[idx] = NULL;
+	pthread_mutex_unlock(&reefs_dirs_lock);
+	return 0;
+}
+
+static int reefs_mkdirs(char *dir, mode_t mode)
+{
+	int i, len = supp_strlen(dir), ret;
+
+	DMSG("mkdirs %s\n", dir);
 
 	for (i = 0; i <= len; i++) {
 		if ((dir[i] == '/' && i > 0) || (i == len)) {
 			dir[i] = 0;
-			if (access(dir, R_OK) < 0) {
-				MSG("mkdirs %s\n", dir);
-
-				ret = mkdir(dir, mode);
-				if (ret < 0) {
-					ret = errno;
-					MSG("mkdir %s, failed %d\n", dir, ret);
-					return -ret;
-				}
+			ret = mkdir(dir, mode);
+			if (ret < 0 && errno != EEXIST && errno != EISDIR) {
+				ret = errno;
+				EMSG("mkdir %s, failed %d\n", dir, ret);
+				return -ret;
 			}
 			dir[i] = '/';
 		}
@@ -57,57 +94,145 @@ static int reefs_mkdirs(const char *path, mode_t mode)
 	return 0;
 }
 
-static void reefs_path_prefix(char *out, const char *in)
+static int reefs_build_path(char *out, const char *in, size_t buflen)
 {
-	snprintf(out, REEFS_PATH_MAX, "%s%s", REEFS_PATH, in);
+	int ret;
+
+	/* Skip leading slashes to normalize path */
+	if (*in == '/')
+		in++;
+
+	/*
+	 * Build the full path: prefix + in.
+	 */
+	ret = supp_strlcpy(out, REEFS_PATH, buflen);
+	ret = supp_strlcat(out, in, buflen);
+	if (ret >= buflen)
+		return -ENAMETOOLONG;
+
+	DMSG("reefs_build_path dst %s ilen %d\n", out, ret);
+	if (!supp_strstr(out, in))
+		EMSG("reefs_build_path FAIL %s\n", in);
+
+	/* Reject path traversal sequences */
+	if (supp_strstr(out, ".."))
+		return -EINVAL;
+
+	/* Reject hidden files starting with . (except ./ prefix) */
+	if (supp_strstr(out, "/.") && !supp_strstr(out, "/.."))
+		return -EINVAL;
+
+	return 0;
 }
 
 static inline int reefs_isroot(const char *path)
 {
-	if ((strlen(path) == 0) ||
-		(strlen(path) == 1 && path[0] == '/'))
+	size_t len = supp_strlen(path);
+
+	if (len == 0 || (len == 1 && path[0] == '/'))
 		return true;
 
 	return false;
 }
 
-static int reefs_open(struct reefs_cmd *r)
+static inline char *reefs_dirname(char *path)
 {
-	int ret = -1, l = 0, i = 0;
-	char path[REEFS_PATH_MAX];
-	char dir[REEFS_PATH_MAX];
+	int i = 0;
+	unsigned int l = 0;
 
-	reefs_path_prefix(path, (const char *)r->data);
+	if (!path || *path == 0)
+		return NULL;
 
-	if (r->flags & O_CREAT) {
-		MSG("creating %s\n", path);
-		/* make parent directory */
-		l = strlen(path);
-		if (l >= REEFS_PATH_MAX)
-			return -ENAMETOOLONG;
-		memcpy(dir, path, l);
-		dir[l] = 0;
+	l = supp_strlen(path);
 
-		for (i = l - 1; i >= 0; i--) {
-			if (dir[i] == '/') {
-				dir[i] = 0;
-				if (i != l - 1)
-					break;
-			}
-		}
-		ret = reefs_mkdirs(dir, 0700);
-		if (ret != 0)
-			return ret;
-	} else {
-		MSG("opening %s\n", path);
+	while (l && (path[l - 1] == '/')) {
+		path[l - 1] = 0;
+		l--;
 	}
 
-	ret = open(path, r->flags, 0600);
+	for (i = l - 1; i >= 0; i--) {
+		if (path[i] == '/') {
+			path[i] = 0;
+			break;
+		}
+	}
+
+	return (i < 0 || *path == 0) ? "/" : path;
+}
+
+static int reefs_flags_from_rpc(int rpc_flags)
+{
+	int flags = 0;
+
+	if ((rpc_flags & 0x3) == REEFS_O_RDONLY)
+		flags |= O_RDONLY;
+	else if ((rpc_flags & 0x3) == REEFS_O_WRONLY)
+		flags |= O_WRONLY;
+	else if ((rpc_flags & 0x3) == REEFS_O_RDWR)
+		flags |= O_RDWR;
+
+	if (rpc_flags & REEFS_O_CREAT)
+		flags |= O_CREAT;
+	if (rpc_flags & REEFS_O_EXCL)
+		flags |= O_EXCL;
+	if (rpc_flags & REEFS_O_TRUNC)
+		flags |= O_TRUNC;
+	if (rpc_flags & REEFS_O_APPEND)
+		flags |= O_APPEND;
+
+	return flags;
+}
+
+static int reefs_whence_from_rpc(int rpc_whence)
+{
+	if (rpc_whence == REEFS_SEEK_SET)
+		return SEEK_SET;
+	if (rpc_whence == REEFS_SEEK_CUR)
+		return SEEK_CUR;
+	if (rpc_whence == REEFS_SEEK_END)
+		return SEEK_END;
+	return -1;
+}
+
+static int reefs_open(struct reefs_cmd *r)
+{
+	int ret = -1;
+	char path[REEFS_PATH_MAX];
+	char dir[REEFS_PATH_MAX];
+	int flags = reefs_flags_from_rpc(r->flags) | O_NOFOLLOW;
+
+	if (reefs_build_path(path, r->data, sizeof(path)) != 0)
+		return -EACCES;
+
+	if (flags & O_CREAT) {
+		DMSG("creating %s\n", path);
+		/* make parent directory */
+		supp_strcpy(dir, path);
+		/*
+		 * Hold reefs_meta_lock across mkdirs + open to prevent a
+		 * concurrent rmdir from removing the parent directory between
+		 * the two operations.
+		 */
+		pthread_mutex_lock(&reefs_meta_lock);
+		ret = reefs_mkdirs(reefs_dirname(dir), 0700);
+		if (ret != 0) {
+			pthread_mutex_unlock(&reefs_meta_lock);
+			return ret;
+		}
+		ret = open(path, flags, 0600);
+		pthread_mutex_unlock(&reefs_meta_lock);
+	} else {
+		DMSG("opening %s\n", path);
+		pthread_mutex_lock(&reefs_meta_lock);
+		ret = open(path, flags, 0600);
+		pthread_mutex_unlock(&reefs_meta_lock);
+	}
+
 	if (ret < 0) {
 		ret = -errno;
-		MSG("open %s, error %d\n", path, errno);
+		DMSG("open %s, errno %d\n", path, -ret);
 	} else {
-		MSG("open %s, fd %d\n", path, ret);
+		DMSG("open %s, fd %d\n", path, ret);
 	}
 
 	return ret;
@@ -119,7 +244,7 @@ static int reefs_close(struct reefs_cmd *r)
 
 	ret = ret ? -errno : ret;
 
-	MSG("close %d, ret %d\n", r->fd, ret);
+	DMSG("close %d, ret %d\n", r->fd, ret);
 
 	return ret;
 }
@@ -135,9 +260,10 @@ static ssize_t reefs_read(struct reefs_cmd *r)
 
 		if (rc < 0) {
 			rc = -errno;
-			MSG("read fail, errno %d\n", errno);
+			EMSG("read %d fail, errno %d\n", r->fd, errno);
 			break;
 		} else if (rc == 0) {
+			DMSG("read %d rc=0, errno %d\n", r->fd, errno);
 			break;
 		}
 
@@ -158,9 +284,10 @@ static ssize_t reefs_write(struct reefs_cmd *r)
 
 		if (rc < 0) {
 			rc = -errno;
-			MSG("write fail, errno %d\n", errno);
+			EMSG("write %d fail, errno %d\n", r->fd, errno);
 			break;
 		} else if (rc == 0) {
+			DMSG("write %d rc=0, errno %d\n", r->fd, errno);
 			break;
 		}
 
@@ -170,7 +297,7 @@ static ssize_t reefs_write(struct reefs_cmd *r)
 	return (rc < 0) ? rc : offset;
 }
 
-static int reefs_truncate(struct reefs_cmd *r)
+static int reefs_ftruncate(struct reefs_cmd *r)
 {
 	int ret = ftruncate(r->fd, r->len);
 
@@ -180,12 +307,12 @@ static int reefs_truncate(struct reefs_cmd *r)
 static int reefs_unlink(struct reefs_cmd *r)
 {
 	int ret = -1;
-
 	char path[REEFS_PATH_MAX];
 
-	reefs_path_prefix(path, (const char *)r->data);
+	if (reefs_build_path(path, r->data, sizeof(path)) != 0)
+		return -EACCES;
 
-	MSG("unlink file %s\n", path);
+	DMSG("unlink file %s\n", path);
 
 	ret = unlink(path);
 
@@ -197,73 +324,92 @@ static int reefs_rename(struct reefs_cmd *r)
 	int ret = -1;
 	char oldpath[REEFS_PATH_MAX];
 	char newpath[REEFS_PATH_MAX];
-	size_t offset = strnlen((const char *)r->data, REEFS_PATH_MAX) + 1;
+	size_t first_len = supp_strnlen(r->data, REEFS_PATH_MAX);
 
-	reefs_path_prefix(oldpath, (const char *)r->data);
-	reefs_path_prefix(newpath, (const char *)r->data + offset);
+	/* Ensure first path is null-terminated within bounds */
+	if (first_len >= REEFS_PATH_MAX)
+		return -ENAMETOOLONG;
 
-	MSG("rename %s -> %s\n", oldpath, newpath);
+	if (reefs_build_path(oldpath, r->data,
+			sizeof(oldpath)) != 0)
+		return -EACCES;
+	if (reefs_build_path(newpath, r->data +
+			first_len + 1, sizeof(newpath)) != 0)
+		return -EACCES;
+
+	DMSG("rename %s -> %s\n", oldpath, newpath);
 
 	if (reefs_isroot(r->data))
 		return -EBUSY;
 
-	if (access(newpath, R_OK) == 0)
-		return -EEXIST;
-
+	pthread_mutex_lock(&reefs_meta_lock);
 	ret = rename(oldpath, newpath);
+	pthread_mutex_unlock(&reefs_meta_lock);
 
 	return ret ? -errno : ret;
 }
 
 static int reefs_mkdir(struct reefs_cmd *r)
 {
+	int ret = -1;
 	char path[REEFS_PATH_MAX];
 
-	reefs_path_prefix(path, (const char *)r->data);
+	if (reefs_build_path(path, r->data, sizeof(path)) != 0)
+		return -EACCES;
 
-	MSG("mkdir %s\n", path);
+	pthread_mutex_lock(&reefs_meta_lock);
+	ret = reefs_mkdirs(path, 0700);
+	pthread_mutex_unlock(&reefs_meta_lock);
 
-	return reefs_mkdirs(path, r->flags);
+	return ret;
 }
 
 static int reefs_opendir(struct reefs_cmd *r)
 {
 	int ret = -1;
+	int dfd = -1;
 	DIR *d = NULL;
-	pthread_key_t k = 0;
 	char path[REEFS_PATH_MAX];
 
-	reefs_path_prefix(path, (const char *)r->data);
+	if (reefs_build_path(path, r->data, sizeof(path)) != 0)
+		return -EACCES;
 
-	MSG("opendir %s\n", path);
+	DMSG("opendir %s\n", path);
 
-	ret = pthread_key_create(&k, NULL);
-	if (ret != 0)
-		return -ENOMEM;
-
-	d = opendir(path);
-	if (d == NULL) {
+	dfd = open(path, O_RDONLY | O_DIRECTORY | O_NOFOLLOW);
+	if (dfd < 0) {
 		ret = -errno;
-		MSG("opendir %s, errno %d\n", path, errno);
-		pthread_key_delete(k);
+		DMSG("opendir %s, errno %d\n", path, -ret);
 		return ret;
 	}
 
-	pthread_setspecific(k, d);
+	d = fdopendir(dfd);
+	if (!d) {
+		ret = -errno;
+		close(dfd);
+		return ret;
+	}
 
-	return k;
+	ret = reefs_alloc_dir(d);
+	if (ret < 0) {
+		closedir(d);
+		return ret;
+	}
+
+	return ret;
 }
 
 static int reefs_closedir(struct reefs_cmd *r)
 {
 	int ret = -1;
-	pthread_key_t k = r->fd;
-	DIR *d = pthread_getspecific(k);
+	DIR *d = reefs_get_dir(r->fd);
 
-	if (d == NULL)
+	if (!d)
 		return -EINVAL;
 
-	pthread_key_delete(k);
+	ret = reefs_free_dir(r->fd);
+	if (ret != 0)
+		return ret;
 
 	ret = closedir(d);
 
@@ -272,13 +418,13 @@ static int reefs_closedir(struct reefs_cmd *r)
 
 static int reefs_readdir(struct reefs_cmd *r)
 {
-	DIR *dir = pthread_getspecific(r->fd);
+	DIR *dir = reefs_get_dir(r->fd);
 	struct reefs_dirent *d = (struct reefs_dirent *)r->data;
 	off_t str_len = 0, retlen = 0, lastdoff = telldir(dir);
 	off_t reclen = -1, structsz = 0, cnt = r->len;
 	struct dirent *e = NULL;
 
-	if (d == NULL || dir == NULL)
+	if (!d || !dir)
 		return -EINVAL;
 
 	structsz = sizeof(d->d_reclen) + sizeof(d->d_off) + sizeof(d->d_type);
@@ -288,7 +434,7 @@ static int reefs_readdir(struct reefs_cmd *r)
 			(e->d_name[1] == 0 || e->d_name[1] == '.'))
 			continue;
 
-		str_len = strlen(e->d_name) + 1;
+		str_len = supp_strlen(e->d_name) + 1;
 		reclen = str_len + structsz;
 		reclen = ((reclen + sizeof(long) - 1) / sizeof(long)) * sizeof(long);
 
@@ -304,30 +450,26 @@ static int reefs_readdir(struct reefs_cmd *r)
 		d->d_off = e->d_off;
 		d->d_type = e->d_type;
 		d->d_reclen = reclen;
-		memcpy(d->d_name, e->d_name, str_len);
+		supp_memcpy(d->d_name, e->d_name, str_len);
 
 		retlen += reclen;
 		cnt -= reclen;
 		d = (void *)d + reclen;
 	}
 
-	/* got nothing */
-	if (retlen == 0)
-		retlen = EOF;
-
 	return retlen;
 }
 
 static int reefs_seekdir(struct reefs_cmd *r)
 {
-	DIR *dir = pthread_getspecific(r->fd);
+	DIR *dir = reefs_get_dir(r->fd);
 
-	if (dir == NULL)
+	if (!dir)
 		return -EINVAL;
 
 	seekdir(dir, r->len);
 
-	return -errno;
+	return 0;
 }
 
 static int reefs_rmdir(struct reefs_cmd *r)
@@ -335,21 +477,32 @@ static int reefs_rmdir(struct reefs_cmd *r)
 	int ret = -1;
 	char path[REEFS_PATH_MAX];
 
-	reefs_path_prefix(path, (const char *)r->data);
+	if (reefs_build_path(path, r->data, sizeof(path)) != 0)
+		return -EACCES;
 
-	MSG("rmdir %s\n", path);
+	DMSG("rmdir %s\n", path);
 
 	if (reefs_isroot(r->data))
 		return -EBUSY;
 
+	pthread_mutex_lock(&reefs_meta_lock);
 	ret = rmdir(path);
+	if (ret != 0)
+		ret = -errno;
+	pthread_mutex_unlock(&reefs_meta_lock);
 
-	return ret ? -errno : ret;
+	return ret;
 }
 
 static off_t reefs_lseek(struct reefs_cmd *r)
 {
-	off_t ret = lseek(r->fd, r->len, r->flags);
+	int whence = reefs_whence_from_rpc(r->flags);
+	off_t ret = -EINVAL;
+
+	if (whence < 0)
+		return ret;
+
+	ret = lseek(r->fd, r->len, whence);
 
 	if (ret < 0)
 		ret = -errno;
@@ -357,15 +510,91 @@ static off_t reefs_lseek(struct reefs_cmd *r)
 	return ret;
 }
 
-int reefs_routine(void *data)
+static ssize_t reefs_pread(struct reefs_cmd *r)
+{
+	void *data = r->data;
+	off_t off = r->flags;
+	off_t rb = 0, rc = -1, offset = 0;
+
+	while (offset < r->len) {
+		rb = r->len - offset;
+		rc = pread(r->fd, data + offset, rb, off + offset);
+
+		if (rc < 0) {
+			rc = -errno;
+			EMSG("pread %d fail, errno %d\n", r->fd, errno);
+			break;
+		} else if (rc == 0) {
+			DMSG("pread %d rc=0, errno %d\n", r->fd, errno);
+			break;
+		}
+
+		offset += rc;
+	}
+
+	return (rc < 0) ? rc : offset;
+}
+
+static ssize_t reefs_pwrite(struct reefs_cmd *r)
+{
+	void *data = r->data;
+	off_t off = r->flags;
+	off_t wb = 0, rc = -1, offset = 0;
+
+	while (offset < r->len) {
+		wb = r->len - offset;
+		rc = pwrite(r->fd, data + offset, wb, off + offset);
+
+		if (rc < 0) {
+			rc = -errno;
+			EMSG("pwrite %d fail, errno %d\n", r->fd, errno);
+			break;
+		} else if (rc == 0) {
+			DMSG("pwrite %d rc=0, errno %d\n", r->fd, errno);
+			break;
+		}
+
+		offset += rc;
+	}
+
+	return (rc < 0) ? rc : offset;
+}
+
+static int reefs_fstat(struct reefs_cmd *r)
 {
 	int ret = -1;
-	struct reefs_cmd *cmd = (struct reefs_cmd *)data;
+	struct stat st;
+	struct reefs_stat *rst = (struct reefs_stat *)r->data;
+	int fd = r->fd;
 
-	if (cmd == NULL)
-		return -EINVAL;
+	if (r->flags & REEFS_O_DIRECTORY) {
+		DIR *d = reefs_get_dir(r->fd);
+		if (!d)
+			return -EINVAL;
+		fd = dirfd(d);
+	}
 
-	switch (cmd->op) {
+	ret = fstat(fd, &st);
+	if (ret == 0) {
+		rst->rst_size = st.st_size;
+		rst->rst_atime = st.st_atime;
+		rst->rst_mtime = st.st_mtime;
+		rst->rst_ctime = st.st_ctime;
+	} else {
+		ret = -errno;
+	}
+
+	return ret;
+}
+
+int reefs_routine(struct reefs_cmd *cmd)
+{
+	int ret = -EINVAL;
+
+	if (!cmd)
+		return ret;
+
+	switch (cmd->hdr.op) {
 	case REEFS_OPEN:
 		ret = reefs_open(cmd);
 		break;
@@ -385,7 +614,7 @@ int reefs_routine(void *data)
 		ret = reefs_rename(cmd);
 		break;
 	case REEFS_TRUNC:
-		ret = reefs_truncate(cmd);
+		ret = reefs_ftruncate(cmd);
 		break;
 	case REEFS_MKDIR:
 		ret = reefs_mkdir(cmd);
@@ -408,11 +637,20 @@ int reefs_routine(void *data)
 	case REEFS_SEEK:
 		ret = reefs_lseek(cmd);
 		break;
+	case REEFS_FSTAT:
+		ret = reefs_fstat(cmd);
+		break;
+	case REEFS_PREAD:
+		ret = reefs_pread(cmd);
+		break;
+	case REEFS_PWRITE:
+		ret = reefs_pwrite(cmd);
+		break;
 	default:
 		ret = -ENOTSUP;
 		break;
 	}
 
-	cmd->ret = ret;
+	cmd->hdr.ret = mbedtee_supp_errno_to_gp(ret);
 	return ret;
 }
